refactor(cast): inline integer bound implementation
